+++ src/index.js
@@ -1,1 +1,2 @@
 export Pager from './pager';
+export FilePicker from './filepicker';