+++ example/components/pager.js
@@ -44,7 +44,7 @@ class PagerTest extends React.Component {
       <section>
         <h5>Pager</h5>
         <p>Pager based on ToolBox components with default button styles. Example: 1.</p>
-
+        
         <div className={style.pager}>
           <Pager
               prevButtonLabel={<FontIcon value='chevron_left' />}
@@ -57,9 +57,11 @@ class PagerTest extends React.Component {
               onPageChange={this.onPageChange1}
           />
         </div>
+        
+        
 
         <p>Pager based on ToolBox components with custom button styles. Example: 2.</p>
-
+        
         <div className={style.pager}>
           <Pager
               prevButtonLabel={<FontIcon value='chevron_left' />}
@@ -74,9 +76,11 @@ class PagerTest extends React.Component {
               onPageChange={this.onPageChange2}
           />
         </div>
+        
+        
 
         <p>Pager based on ToolBox components with custom button styles. Example: 3.</p>
-
+        
         <div className={style.pager}>
           <Pager
               prevButtonLabel={<FontIcon value='chevron_left' />}
@@ -86,15 +90,18 @@ class PagerTest extends React.Component {
               leftRightArrowButtonStyles={{ primary: true, raised: true }}
               leftRightRangeButtonStyles={{ raised: true}}
               pagesButtonStyles={{raised: true }}
+              firstLastPagesButtonStyles={{ raised: true}}
               totalPages={this.state.totalPages}
               currentPage={this.state.currentPage}
               visiblePagesBlockSize={this.props.visiblePagesBlockSize}
               onPageChange={this.onPageChange2}
           />
         </div>
+        
+        
 
         <p>Pager based on ToolBox components with custom theme. Example: 4.</p>
-
+        
         <div className={style.pager}>
           <Pager
               prevButtonLabel={<FontIcon value='chevron_left' />}
@@ -104,6 +111,7 @@ class PagerTest extends React.Component {
               leftRightArrowButtonStyles={{ primary: true, raised: true }}
               leftRightRangeButtonStyles={{ raised: true}}
               pagesButtonStyles={{raised: true }}
+              firstLastPagesButtonStyles={{ raised: true}}
               totalPages={this.state.totalPages}
               currentPage={this.state.currentPage}
               visiblePagesBlockSize={this.props.visiblePagesBlockSize}
@@ -111,6 +119,28 @@ class PagerTest extends React.Component {
               theme={style}
           />
         </div>
+
+        
+        
+
+        <p>Pager based on ToolBox components with custom theme. Example: 5.</p>
+        
+        <div className={style.pager}>
+          <Pager
+              prevButtonLabel={<FontIcon value='chevron_left' />}
+              nextButtonLabel={<FontIcon value='chevron_right' />}
+              rangeLeftButtonLabel={<FontIcon value='more_horiz' />}
+              rangeRightButtonLabel={<FontIcon value='more_horiz' />}
+              leftRightArrowButtonStyles={{ primary: true}}
+              firstLastPagesButtonStyles={{ raised: true}}
+              totalPages={this.state.totalPages}
+              currentPage={this.state.currentPage}
+              visiblePagesBlockSize={this.props.visiblePagesBlockSize}
+              onPageChange={this.onPageChange2}
+              theme={style}
+          />
+        </div>
+        
 
         <Input
           type='number'